requirements.txt
# 网络工具箱 - Python依赖
# 使用方法: pip install -r requirements.txt

# GUI框架
ttkbootstrap>=1.10.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# SSH采集
paramiko>=3.0.0
pysocks>=1.7.0

# 网络设备解析
net-inspect>=0.3.0
textfsm>=1.1.0

# 配置模板
jinja2>=3.0.0

# 拓扑图生成
graphviz>=0.20.0
pyvis>=0.3.0
